Match BlockRemoveNeverReformSingleBlock

+++ include/block.h
@@ -11,7 +11,7 @@ u8 BlockDestroySingleBreakableBlock(struct ClipdataBlockData* pClipBlock);
 u8 BlockDestroySquareBlock(struct ClipdataBlockData* pClipBlock);
 u32 BlockStoreSingleNeverReformBlock(u16 xPosition, u16 yPosition);
 void BlockRemoveNeverReformBlocks(void);
-void BlockRemoveNeverReformSingleBlock(u16 xPosition, u16 yPosition);
+void BlockRemoveNeverReformSingleBlock(u8 xPosition, u8 yPosition);
 void BlockShiftNeverReformBlocks(void);
 u32 BlockCheckRevealOrDestroyNonBombBlock(struct ClipdataBlockData* pClipBlock);
 u32 BlockCheckRevealOrDestroyBombBlock(struct ClipdataBlockData* pClipBlock);
